requirements.txt
PySide6>=6.6
qasync>=0.27
httpx[http2,socks]>=0.27
pydantic>=2.5
PyNaCl>=1.5
DrissionPage>=4.1
playwright>=1.40
pymailtm>=1.1
pyperclip>=1.9
psutil>=5.9
websockets>=12.0
aiohttp>=3.9
loguru>=0.7
pywin32>=306
tqdm>=4.66
rich>=13.8
pyinstaller>=6.11
